% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{biggest_subnetwork}
\alias{biggest_subnetwork}
\title{find biggest subnetwork}
\usage{
biggest_subnetwork(edgelist)
}
\arguments{
\item{edgelist}{An edge list in dataframe format.}
}
\value{
An edge list in dataframe format.
}
\description{
Find and return the biggest (in terms of number of edges) connected subnetwork
of a given network.
}
\details{
Uses \code{\link{colour_network}} to find isolated subnetworks, then returns only
the edges of the subnetwork with the highest number of edges.
}
