% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{children}
\alias{children}
\title{children of a node}
\usage{
children(edgelist, node)
}
\arguments{
\item{edgelist}{A network in edgelist format.}

\item{node}{A single node id or list of node ids.}
}
\value{
An array of nodes or a list of arrays of nodes (if no node was provided).
}
\description{
Find all direct children of a node.
}
\details{
Find and return all children for a given node (or list of nodes) in a network. If 
no node is provided the function is applied to all nodes in the network.
}
\examples{
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"))
# get children of specific nodes
children(el, "A")
children(el, c("A", "B"))
# apply to all nodes in the network
children(el)
}
