% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colour_network}
\alias{colour_network}
\title{colour_network}
\usage{
colour_network(edge_list)
}
\arguments{
\item{edge_list}{A dataframe containing a list of edges (see \code{\link{popsnetwork}}
for a description of possible formats).}
}
\value{
An integer vector with the sub-network id of each edge. Note that id's start at
1 and are not guaranteed to be contiguous.
}
\description{
Identify separate sub-networks.
}
\details{
This function identifies completely separate sub-networks in a network
described as an edge list.
}
