% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{colourNetwork}
\alias{colourNetwork}
\title{colourNetwork}
\usage{
colourNetwork(edgeList)
}
\arguments{
\item{edgeList}{A dataframe containing a list of edges.}
}
\value{
An integer vector with the sub-network id of each edge. Note that id's start at
1 and are not guaranteed to be contiguous.
}
\description{
Identify separate sub-networks.
}
\details{
This function identifies completely separate sub-networks in a network
described as an edge list.
}
