% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_EHamming}
\alias{distances_EHamming}
\title{distances_EHamming}
\usage{
distances_EHamming(p_net, skip_empty = TRUE)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{skip_empty}{Whether to return NA for empty nodes.}
}
\value{
A matrix of all distances.
}
\description{
Calculate genetic distances within a network as expected values of Hamming
distances.
}
\details{
This function calculates the genetic distance of all pairs of nodes in a 
network by calculating per pair of nodes the average Hamming distance between them 
(more precisely the expected value of the Hamming distance between two individuals 
randomly selected from each of the nodes).
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# simulate
res <- popgen_dirichlet(net, 0.3, ini_freqs)

# get distances
distances_EHamming(res)
}
