% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_freqdist}
\alias{distances_freqdist}
\title{distances_freqdist}
\usage{
distances_freqdist(p_net, skip_empty = TRUE)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{skip_empty}{Whether to return NA for empty nodes.}
}
\value{
A matrix of all distances.
}
\description{
Calculate genetic dissimilarities within a network.
}
\details{
This function calculates the dissimilarity (mean square distance in
allele frequencies) of all pairs of nodes in a network.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# simulate
res <- popgen_dirichlet(net, 0.3, ini_freqs)

# get distances
distances_freqdist(res)
}
