% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_net}
\alias{distances_net}
\title{distances_net}
\usage{
distances_net(p_net)
}
\arguments{
\item{p_net}{A popsnetwork object.}
}
\value{
A matrix of all distances.
}
\description{
Calculate genetic distances within a network.
}
\details{
This function calculates the genetic distance between all pairs of nodes in 
a popsnetwork.
}
