% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_topology}
\alias{distances_topology}
\title{distances_topology}
\usage{
distances_topology(p_net)
}
\arguments{
\item{p_net}{A popsnetwork object.}
}
\value{
A matrix with distance values.
}
\description{
Calculate topological distances between nodes in a network.
}
\details{
This function calculates the topological distance (number of edges in
the shortest path) between all pairs of nodes in a network.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# get distances
distances_topology(net)
}
