% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_alleles.popsnetwork}
\alias{draw_alleles.popsnetwork}
\title{draw_alleles.popsnetwork}
\usage{
\method{draw_alleles}{popsnetwork}(p_net, nodes, n = 1L)
}
\arguments{
\item{p_net}{a PopsNet object.}

\item{nodes}{A vector of node ids (either integer or factor).}

\item{n}{How many alleles to draw per node.}
}
\value{
A dataframe with one column per node containing a list of allele ids.
}
\description{
Draw a set of alleles from the network.
}
\details{
Draw a random set of alleles from a number of nodes in the network. This will
*only* work if allele frequencies have been set or simulated.
}
