% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_alleles.popsnode}
\alias{draw_alleles.popsnode}
\title{draw_alleles.popsnode}
\usage{
\method{draw_alleles}{popsnode}(p_node, n)
}
\arguments{
\item{p_node}{A popsnode object.}

\item{n}{Number of alleles to draw.}
}
\value{
A vector of allele ids.
}
\description{
Draw a set of alleles from a single node.
}
\details{
Draw a random set of alleles from a given node. This will only work if
allele frequencies have been set (manually or by simulation).
}
