% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_alleles}
\alias{draw_alleles}
\title{draw_alleles}
\usage{
draw_alleles(p_net, nodes, n = 1L)
}
\arguments{
\item{p_net}{a PopsNet object.}

\item{nodes}{A vector of node ids (either integer or factor).}

\item{n}{How many alleles to draw per node.}
}
\value{
A dataframe with one column per node containing a list of allele ids.
}
\description{
Draw a set of alleles from the network.
}
\details{
Draw a random set of alleles from a number of nodes in the network. This will
*only* work if allele frequencies have been set or simulated.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# simulate
res <- popgen_dirichlet(net, 0.3, ini_freqs)

# get some data
draw_alleles(res, as.factor(c("C", "D")))
}
