% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_isolates.popsnode}
\alias{draw_isolates.popsnode}
\title{draw_isolates.popsnode}
\usage{
\method{draw_isolates}{popsnode}(p_node, n)
}
\arguments{
\item{p_node}{A popsnode object.}

\item{n}{Number of isolates to draw.}
}
\value{
A vector of allele counts.
}
\description{
Draw a set of isolates from a single node.
}
\details{
Draw a random set of isolates from a given node. This will only work if
allele frequencies have been set (manually or by simulation).
}
