% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_isolates}
\alias{draw_isolates}
\title{draw_isolates}
\usage{
draw_isolates(p_net, samples)
}
\arguments{
\item{p_net}{a PopsNet object.}

\item{samples}{Number of samples to draw from each node. This has to be a dataframe
with node ids (see \code{\link{popsnetwork}}) in the first and number of isolates to 
draw in the second column.}
}
\value{
A dataframe containing node id in $node and number of isolates with allele \code{x}
in $\code{allele_x}.
}
\description{
Draw a set of isolates from the network.
}
\details{
Draw a random set of isolates from a number of nodes in the network. This will
*only* work if allele frequencies have been set or simulated.
}
