% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edgeList}
\alias{edgeList}
\title{egdeList}
\usage{
edgeList(pNet)
}
\arguments{
\item{pNet}{A PopsNet object.}
}
\value{
A dataframe with from, to, rates and rates_infected.
}
\description{
Get a list of edges in a dataframe.

This function returns a list of the edges in the network in a format
that is suitable for plotting with e.g. iGraph.
}
