% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{epic_to_edge_list}
\alias{epic_to_edge_list}
\title{epicontact to edge list}
\usage{
epic_to_edge_list(x)
}
\arguments{
\item{x}{An \code{epicontacts} object.}
}
\description{
Convert epicontacts to edgelist.
}
\details{
Extract an edgelist suitable for popsnetwork from an epicontacts object.
Note that this will set all rates to 1.
}
\examples{
if (require("outbreaks") && require("epicontacts")) {
 x <- make_epicontacts(linelist = mers_korea_2015$linelist,
                       contacts = mers_korea_2015$contacts)
 edgelist <- epic_to_edge_list(x)
}
}
