% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{epicToEdgelist}
\alias{epicToEdgelist}
\title{epicToEdgelist}
\usage{
epicToEdgelist(x)
}
\arguments{
\item{x}{An \code{epicontacts} object.}
}
\details{
Extract an edgelist suitable for PopsNetwork from an epicontacts object.
Note that this will set all rates to 1.
}
\examples{
if (require("outbreaks") && require("epicontacts")) {
 x <- make_epicontacts(linelist = mers_korea_2015$linelist,
                       contacts = mers_korea_2015$contacts)
 edgelist <- epicToEdgelist(x)
}
}
