% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{generate_PA}
\alias{generate_PA}
\title{generate_PA}
\usage{
generate_PA(n_nodes, n_sources, m_dist, zero_appeal = 1, compact = TRUE)
}
\arguments{
\item{n_nodes}{Number of (non-source) nodes to generate.}

\item{n_sources}{Number of source nodes to initialize the network with (has to be at least
1). Note that there is no guarantee all source nodes will become part of the network.}

\item{m_dist}{The probability distribution to draw the number of inputs for new nodes
from. m_dist will be normalized, therefore it does not have to sum up to 1.}

\item{zero_appeal}{Constant to be added to the nodes' attractiveness.}

\item{compact}{Whether to remove isolated source nodes.}
}
\value{
An edgelist as a dataframe.
}
\description{
Generate a random transport network using preferential attachment.
}
\details{
This function generates a random scale-free network. It uses the Barabasi-Albert 
preferential attachment algorithm, slightly modified to allow for directedness and 
isolated initial nodes.
}
