% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_popsnode}
\alias{get_popsnode}
\title{get_popsnode}
\usage{
get_popsnode(p_net, id)
}
\arguments{
\item{p_net}{A PopsNet object.}

\item{id}{The id of the node to return.}
}
\value{
A popsnode object.
}
\description{
Pick a single node from the network.
}
