% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getPopsNode}
\alias{getPopsNode}
\title{getPopsNode}
\usage{
getPopsNode(pNet, id)
}
\arguments{
\item{pNet}{A PopsNet object.}

\item{id}{The id of the node to return.}
}
\value{
A PopsNode object.
}
\description{
Pick a single node from the network.
}
