% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nodeList}
\alias{nodeList}
\title{nodeList}
\usage{
nodeList(pNet)
}
\arguments{
\item{pNet}{A PopsNet object.}
}
\value{
A dataframe with id and rate_infected.
}
\description{
Get a list of nodes in a dataframe.

This function returns a list of the nodes in the network in a format
that is suitable for plotting with e.g. iGraph.
}
