% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{nodes}
\alias{nodes}
\title{nodes}
\usage{
nodes(edgelist)
}
\arguments{
\item{edgelist}{An edgelist as two columns of node ids (from, to).}
}
\value{
A list or vector of node ids.
}
\description{
return a list of nodes
}
\details{
Returns a list of node ids for a graph.
}
