% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{path_distances}
\alias{path_distances}
\title{path_distances}
\usage{
path_distances(net)
}
\arguments{
\item{net}{A popsnetwork object.}
}
\value{
The distances in a matrix (see \code{\link{igraph::distances}}).
}
\description{
Obtain the shortest distances between all pairs of nodes.
}
\details{
Uses \code{\link{igraph::distances}} to calculate the shortest path for
all pairs of nodes in the network.
}
