% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{perfect_binary}
\alias{perfect_binary}
\title{perfect_binary}
\usage{
perfect_binary(size)
}
\arguments{
\item{size}{Size of the tree in number of edges from the root to any leaf (which is
equal to depth-1).}
}
\value{
An edge list in dataframe format.
}
\description{
Create a perfect binary tree of a given size.
}
\details{
Create a perfect, full binary tree with a given depth.
}
