% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.popsnetwork}
\alias{plot.popsnetwork}
\title{plot.popsnetwork}
\usage{
\method{plot}{popsnetwork}(x, ...)
}
\arguments{
\item{x}{A popsnetwork object.}

\item{...}{Ignored for now.}
}
\description{
Plot a popsnetwork object.
}
\details{
This function will plot a popsnetwork object as a simple graph (using 
\code{\link[igraph]{plot.igraph}}).
}
