% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popgen_dirichlet}
\alias{popgen_dirichlet}
\title{run dirichlet model}
\usage{
popgen_dirichlet(p_net, theta, ini_dist = NULL)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{theta}{Scale parameter of the Dirichlet distribution. At each node the Dirichlet
distribution the new allele frequencies ar drawn from is parameterized by the old
frequencies multiplied by theta.}

\item{ini_dist}{Initial distribution of allele frequencies (optional). ini_dist has to be 
a list
containing a vector of node IDs (see \code{\link{popsnetwork}}) and
a matrix of allele frequencies. Note that *any* node pre-set in this
way will effectively be treated as a source and hide nodes that are further upstream.}
}
\value{
A new popsnetwork object with allele frequencies set for each node.
}
\description{
Simulate spread of pathogens on the network using a Dirichlet
distribution to approximate genetic drift.
}
\details{
This function simulates the change of gene frequencies in a population
of pathogens as they spread through the transport network starting at the 
external sources (see \code{\link{popsnetwork}}). At each node founder
effects are assumed to change composition of the population. This change is 
approximated by drawing a set of allele frequencies from a Dirichlet distribution.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# run simulation on initialized net
ini_net <- set_allele_freqs(net, ini_freqs)
popgen_dirichlet(ini_net, 0.3)

# or we can initialize and run in one call
popgen_dirichlet(net, 0.3, ini_freqs)
}
