% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popsnetwork}
\alias{popsnetwork}
\title{popsnetwork}
\usage{
popsnetwork(links, external, transmission = 0)
}
\arguments{
\item{links}{A dataframe describing all edges in the graph in the columns 
$inputs, $outputs and $rates. Note that inputs and outputs refer to node ids.}

\item{external}{A dataframe describing external inputs into the network as $nodes
and $rates.}

\item{transmission}{Rate of infection within nodes.}
}
\value{
A popsnetwork object.
}
\description{
Create a popsnetwork object.
}
\details{
A popsnetwork object stores the nodes and edges making up a food transport
network and associated data describing spread of infected diseases on the
network. A network is created from a tabular description of its edges.
}
