% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.popsnetwork}
\alias{print.popsnetwork}
\title{print.popsnetwork}
\usage{
\method{print}{popsnetwork}(x, ...)
}
\arguments{
\item{x}{A popsnetwork object.}

\item{...}{Ignored for now.}
}
\description{
Print a popsnetwork object.
}
\details{
This will print a textual summary of a popsnetwork object. It will produce two 
tables: 
\itemize{ 
\item A list of all nodes with id (in the same format as provided in the constructor), 
proportion
of infected material present, overall input rate and genetic composition. 
\item A list of all edges with ids of connected nodes, transfer rate and proportion of
infected material being transferred.}
}
