% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpathsonpaths.R
\docType{package}
\name{rpathsonpaths}
\alias{rpathsonpaths}
\alias{rpathsonpaths-package}
\title{rpathsonpaths: A package for simulating the spread of pathogens on food transport networks.}
\description{
rpathsonpaths simulates the spread of pathogens on arbitrary transport networks on the
genetic level. Given a network and a set of introduction points it is possible to quickly
generate the genetic setup and the proportion of infected units in the leaf nodes of the 
network.
}
