% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{rpop_to_dibbler}
\alias{rpop_to_dibbler}
\title{rpop_to_dibbler}
\usage{
rpop_to_dibbler(network)
}
\arguments{
\item{network}{A popsnetwork object.}
}
\value{
A dibbler object.
}
\description{
convert an rpopsnetwork object to a dibbler object
}
\details{
This function converts a popsnetwork object into a dibbler object.
}
