% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{run_popsnet}
\alias{run_popsnet}
\title{run_popsnet}
\usage{
run_popsnet(edgelist, ini_input, ini_infd, ini_freqs, n = 1L,
  transmission = 0, decay = -1, theta = 1, spread_model = "units",
  drift_model = "units", checks = FALSE)
}
\arguments{
\item{edgelist}{The list of edges as a dataframe, with origin in the first and target in 
the second column. If a third column is present it is interpreted as transfer rates per
edge.}

\item{ini_input}{Net input rates per source node. Note that source nodes in the network
will be automatically detected and will be processed in order of node id. ini_input can 
either be a single number that will be a repeated for each source node or a list containing
one element per source node.}

\item{ini_infd}{Input rate of infected units per source nodes. Format requirements are
identical to those for \code{ini_input}.}

\item{ini_freqs}{Allele frequencies per source node. This has to be either a matrix with
one row per source node or a list of initialization objects (each of them containing
a vector of node ids and a matrix of frequencies, see \code{\link{set_allele_freqs}}). In
the latter case n simulations are run per initialization.}

\item{n}{Number of simulation runs to perform.}

\item{transmission}{Rate of infection in nodes.}

\item{decay}{Decay rate in nodes used to infer transfer rates. If this is negative all 
transfer rates will be assumed to be 1 (which in most cases is not a very useful value)
unless edgelist has a third column (see above). If decay is not negative transfer rates
will be inferred assuming preservation of mass with decay at the nodes. Note that the ibm
simulation requires output to be smaller or equal to input for all nodes.}

\item{theta}{The shape parameter for the Dirichlet distribution.}

\item{spread_model}{Either "fluid" or "units".}

\item{drift_model}{Either "dirichlet" or "units".}

\item{checks}{Whether to perform some sanity checks on the graph before simulating (slow).}
}
\value{
A list containing the result of the simulation(s) as first and the raw network as
the second element.
}
\description{
Create a popsnetwork and run the transport simulation.
}
\details{
This function creates a popsnetwork object from the given parameters and runs 
either the ibm or dirichlet simulation on it.
}
