% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_allele_freqs}
\alias{set_allele_freqs}
\title{set allele frequencies}
\usage{
set_allele_freqs(p_net, ini_dist)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{ini_dist}{Initial distribution of allele frequencies. ini_dist has to be 
a list
containing a vector of node IDs (see \code{\link{popsnetwork}}) and
a matrix of allele frequencies. Note that if non-root nodes are initialized in this way
they will be marked as blocked, i.e. they will not receive input in genetic material 
(but will produce output themselves).}
}
\value{
A new popsnetwork object.
}
\description{
Pre-set allele frequencies for some nodes.
}
\details{
Use this function to initialize allele frequencies for some nodes of the network.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)
# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
set_allele_freqs(net, list(as.factor(c("A", "C")), freqs))
}
