% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setAlleleFreqs}
\alias{setAlleleFreqs}
\title{setAlleleFreqs}
\usage{
setAlleleFreqs(pNet, iniDist)
}
\arguments{
\item{pNet}{A PopsNetwork object.}

\item{iniDist}{Initial distribution of allele frequencies. iniDist has to be 
a list
containing a vector of node IDs (@seealso \code{\link{PopsNetwork}}) as $nodes and
a matrix of allele frequencies as $frequencies. Note that *any* node pre-set in this
way will be treated as a source only by spreadDirichlet.}
}
\value{
A new PopsNetwork object.
}
\description{
Pre-set allele frequencies for some nodes.
}
\details{
Use this function to initialize allele frequencies for some nodes of the network.
}
