% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sinks}
\alias{sinks}
\title{get sink nodes}
\usage{
sinks(edge_list)
}
\arguments{
\item{edge_list}{A dataframe containing a list of edges (see \code{\link{popsnetwork}}
for a description of possible formats).}
}
\value{
A list of ids of all sink nodes in the network.
}
\description{
Get a list of sink (i.e. leave) nodes from a network.
}
\details{
Extract a list of sink nodes, i.e. nodes that do not have any
outputs, from a network.
}
\examples{
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"))
sinks(el)
}
