% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SNP_distance}
\alias{SNP_distance}
\title{SNP_distance}
\usage{
SNP_distance(g1, g2)
}
\arguments{
\item{g1}{Sequence 1}

\item{g2}{Sequence 2}
}
\description{
Calculate distance (in number of different SNPs) between two genotypes 
encoded as integers.
}
\details{
This function calculates the difference between two genetic sequences. A sequence
is represented as an integer number where the binary of that number corresponds
to a list of alleles. The distance between to sequences is the number of positions at
which they have different alleles.
}
