% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SNP_distance_pop}
\alias{SNP_distance_pop}
\title{SNP_distance_pop}
\usage{
SNP_distance_pop(p1, p2)
}
\arguments{
\item{p1}{Population 1.}

\item{p2}{Population 2.}
}
\value{
The distance between p1 and p2.
}
\description{
Calculate genetic distance between two populations.
}
\details{
This function calculates the genetic distance between two populations as the
average distance between all pairs of individuals of the two populations.
}
