% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SNP_pop_distance}
\alias{SNP_pop_distance}
\title{SNP_pop_distance}
\usage{
SNP_pop_distance(p1, p2)
}
\arguments{
\item{p1}{Population 1.}

\item{p2}{Population 2.}
}
\description{
Calculate genetic distance between two populations.
}
\details{
This function calculates the genetic distance between two populations as the
average distance between all pairs of individuals of the two populations.
}
