% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spread_dirichlet}
\alias{spread_dirichlet}
\title{spread_dirichlet}
\usage{
spread_dirichlet(p_net, theta, ini_dist = NULL)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{theta}{Scale parameter of the Dirichlet distribution. At each node the Dirichlet
distribution the new allele frequencies ar drawn from is parameterized by the old
frequencies multiplied by theta.}

\item{ini_dist}{Initial distribution of allele frequencies (optional). ini_dist has to be 
a list
containing a vector of node IDs (see \code{\link{popsnetwork}}) and
a matrix of allele frequencies. Note that *any* node pre-set in this
way will effectively be treated as a source and hide nodes that are further upstream.}
}
\value{
A new popsnetwork object with allele frequencies set for each node.
}
\description{
Simulate spread of pathogens on the network using a Dirichlet
distribution to approximate genetic drift.
}
\details{
This function simulates the change of gene frequencies in a population
of pathogens as they spread through the transport network starting at the 
external sources (see \code{\link{popsnetwork}}). At each node founder
effects are assumed to change composition of the population. This change is 
approximated by drawing a set of allele frequencies from a Dirichlet distribution.
}
