% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{depth}
\alias{depth}
\title{depth of a node}
\usage{
depth(edgelist, node)
}
\arguments{
\item{edgelist}{The network in edgelist format.}

\item{node}{A node id or a list of node ids.}
}
\value{
The depth as an integer or an array of integers (if no node was provided).
}
\description{
Calculate depth of a node or network.
}
\details{
Calculate the depth of a node (or list of nodes) in a given network. The depth is
here defined as the length of the shortest path that connects any of the given nodes to
any root node. If no node is provided the function is applied to all nodes in the network.
}
\examples{
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"))
# get min depth of specific nodes
depth(el, "A")
depth(el, c("A", "B"))
# apply to all nodes in the network
# returns a vector, not a number
depth(el)
# to get the network's depth apply to the list of nodes
depth(el, nodes(el))
}
