% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_freqdist}
\alias{distances_freqdist}
\title{distances_freqdist}
\usage{
distances_freqdist(p_net)
}
\arguments{
\item{p_net}{A popsnetwork object.}
}
\value{
A matrix of all distances.
}
\description{
Calculate genetic dissimilarities within a network.
}
\details{
This function calculates the dissimilarity (mean square distance in
allele frequencies) of all pairs of nodes in a network.
}
