% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distances_sample}
\alias{distances_sample}
\title{distances_sample}
\usage{
distances_sample(p_net, n = 1L, skip_empty = TRUE)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{n}{How many samples per node to use for comparison.}

\item{skip_empty}{Whether to return NA for empty nodes.}
}
\value{
A matrix of all distances.
}
\description{
Calculate genetic distances within a network.
}
\details{
This function calculates the genetic distance of all pairs of nodes in a 
network by comparing a number of random samples from each node (using Hamming distance).
}
