% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{draw_isolates}
\alias{draw_isolates}
\title{draw_isolates}
\usage{
draw_isolates(p_net, samples, aggregate = TRUE)
}
\arguments{
\item{p_net}{a PopsNet object.}

\item{samples}{Number of samples to draw from each node. This has to be a dataframe
with node ids (see \code{\link{popsnetwork}}) in the first and number of isolates to 
draw in the second column.}

\item{aggregate}{Whether to return one line per sample taken or to sum up allele counts 
per node.}
}
\value{
A dataframe containing 
\itemize{ 
\item if aggregate==TRUE: node id in \code{$node} and number of isolates with allele 
	\code{x} in \code{$allele_x}.
\item if aggregate==FALSE: node id in \code{$node} and }
}
\description{
Draw a set of isolates from the network.
}
\details{
Draw a random set of isolates from a number of nodes in the network. This will
*only* work if allele frequencies have been set or simulated.
}
\examples{
# create network
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"), rates=c(1.5, 1, 3))
ext <- data.frame(node=c("A", "B"), rate=c(0.3, 0.1))
net <- popsnetwork(el, ext)

# set allele frequencies (2 nodes, 3 alleles)
freqs <- matrix(c(0.1, 0.5, 0.4, 0.9, 0.1, 0), nrow=2, ncol=3, byrow=TRUE)
ini_freqs <- list(as.factor(c("A", "C")), freqs)

# simulate
res <- popgen_dirichlet(net, 0.3, ini_freqs)

# get some data
draw_isolates(res, data.frame(c("C", "D"), c(10, 10)))
}
