% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{edge_list}
\alias{edge_list}
\title{egdeList}
\usage{
edge_list(p_net)
}
\arguments{
\item{p_net}{A PopsNet object.}
}
\value{
A dataframe with from, to, rates and rates_infected.
}
\description{
Get a list of edges in a dataframe.
}
\details{
This function returns a list of the edges in the network in a format
that is suitable for plotting with e.g. iGraph.
}
