% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mutations}
\alias{mutations}
\title{mutations}
\usage{
mutations(edgelist, n_alleles, freq_mutant, n_muts)
}
\arguments{
\item{edgelist}{Two columns of nodes (from, to).}

\item{n_alleles}{How many alleles to initialize the frequency vector to.}

\item{freq_mutant}{Frequency of the mutant allele. Note that the wildtype (allele 0)
will have frequency 1-freq_mutant.}

\item{n_muts}{How many mutations to generate.}
}
\value{
Produces a list of mutations. Each mutation consists of a list of node ids and
a matrix containing allele frequencies for each node.
}
\description{
generate a list of random mutations on a given graph
}
\details{
This function will generate a list of mutations suitable as gene frequencies
with e.g. \code{\link{set_allele_freqs}} or \code{\link{run_popsnet}}. Each mutation 
will be assigned will be assigned to a random node. Allele frequencies will be initialized
for the wild type (0) and the mutant (1) and set to 0 otherwise.
}
