% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{parents}
\alias{parents}
\title{parents}
\usage{
parents(edgelist, node)
}
\arguments{
\item{edgelist}{A network in edgelist format.}

\item{node}{A single node id or list of node ids.}
}
\value{
An array of nodes or a list of arrays of nodes (if no node was provided).
}
\description{
Find all direct parents of a node.
}
\details{
Find and return all parents for a given node (or list of nodes) in a network. If 
no node is provided the function is applied to all nodes in the network.
}
