% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.popsnetwork}
\alias{plot.popsnetwork}
\title{plot.popsnetwork}
\usage{
\method{plot}{popsnetwork}(x, ...)
}
\arguments{
\item{x}{A popsnetwork object.}

\item{...}{Ignored for now.}
}
\description{
Plot a popsnetwork object.
}
\details{
This function will plot a popsnetwork object as a simple graph (using 
\code{\link[igraph]{plot.igraph}}).
}
\examples{
from <- c(0L, 0L, 1L, 2L, 3L, 1L, 5L)
to <- c(1L, 2L, 3L, 3L, 4L, 4L, 2L)
rates <- c(1, 1.5, 0.5, 0.1, 1, 0.1, 0.5)
edgelist <- data.frame(from, to, rates)
ext <- data.frame(c(0L, 5L), c(0.5, 0.5))
net <- popsnetwork(edgelist, ext)
plot(net)
}
