% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popgen_ibm_mixed}
\alias{popgen_ibm_mixed}
\title{popgen_ibm_mixed}
\usage{
popgen_ibm_mixed(p_net, ini_dist = NULL)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{ini_dist}{Initial distribution of allele frequencies (optional). ini_dist has to be 
a list
containing a vector of node IDs (see \code{\link{popsnetwork}}) and
a matrix of allele frequencies. Note that only root nodes can be initialized
in this way.}
}
\value{
A new popsnetwork object with allele frequencies set for each node.
}
\description{
Simulate spread of pathogens on the network using a (very) simple individual-based
model.
}
\details{
This function simulates the change of gene frequencies in a population
of pathogens as they spread through the transport network starting at the 
external sources (see \code{\link{popsnetwork}}). At each node founder
effects are assumed to change composition of the population. This change is 
simulated by directly drawing from the distribution of genotypes and unfected units,
respectively.
}
