% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{popsnetwork}
\alias{popsnetwork}
\title{popsnetwork}
\usage{
popsnetwork(links, external, transmission = 0, checks = FALSE)
}
\arguments{
\item{links}{A dataframe describing all edges in the graph as well as transfer rates
between them. The first two columns are read as inputs and outputs. If there are only 
two columns all rates are assumed to be 1.}

\item{external}{A dataframe describing external inputs into the network as $nodes
and $rates.}

\item{transmission}{Rate of infection within nodes.}

\item{checks}{Perform some basic integrity checks on input data (currently looks for cycles
and disconnected sub-networks).}
}
\value{
A popsnetwork object.
}
\description{
Create a popsnetwork object.
}
\details{
A popsnetwork object stores the nodes and edges making up a food transport
network and associated data describing spread of infected diseases on the
network. A network is created from a tabular description of its edges.
}
\section{Edge lists and node ids}{

Many functions in rpathsonpaths take edge lists as an argument. An edge list is a 
data frame with at least two columns. These columns can be either integer vectors
which will be interpreted as \emph{0-based node indices} or factors. In general where
applicable functions will return the same format they received. Note that formats 
can not be mixed easily. Function arguments (and data frame columns) have to have the 
same type. Furthermore nodes in a \code{popsnetwork} object generated with integer
indices can not be referenced by name.

In general integer indices are considerably faster than factors.
}

