% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.popsnetwork}
\alias{print.popsnetwork}
\title{print popsnetwork}
\usage{
\method{print}{popsnetwork}(x, ...)
}
\arguments{
\item{x}{A popsnetwork object.}

\item{...}{Ignored for now.}
}
\description{
Print a popsnetwork object.
}
\details{
This will print a textual summary of a popsnetwork object. It will produce two 
tables: 
\itemize{ 
\item A list of all nodes with id (in the same format as provided in the constructor), 
proportion
of infected material present, overall input rate and genetic composition. 
\item A list of all edges with ids of connected nodes, transfer rate and proportion of
infected material being transferred.}
}
\examples{
from <- c(0L, 0L, 1L, 2L, 3L, 1L, 5L)
to <- c(1L, 2L, 3L, 3L, 4L, 4L, 2L)
rates <- c(1, 1.5, 0.5, 0.1, 1, 0.1, 0.5)
edgelist <- data.frame(from, to, rates)
ext <- data.frame(c(0L, 5L), c(0.5, 0.5))
net <- popsnetwork(edgelist, ext)
print(net)
}
