% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{set_allele_freqs}
\alias{set_allele_freqs}
\title{set_allele_freqs}
\usage{
set_allele_freqs(p_net, ini_dist)
}
\arguments{
\item{p_net}{A popsnetwork object.}

\item{ini_dist}{Initial distribution of allele frequencies. ini_dist has to be 
a list
containing a vector of node IDs (see \code{\link{popsnetwork}}) and
a matrix of allele frequencies. Note that if non-root nodes are initialized in this way
they will be marked as blocked, i.e. they will not receive input in genetic material 
(but will produce output themselves).}
}
\value{
A new popsnetwork object.
}
\description{
Pre-set allele frequencies for some nodes.
}
\details{
Use this function to initialize allele frequencies for some nodes of the network.
}
