% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sources}
\alias{sources}
\title{sources}
\usage{
sources(edge_list)
}
\arguments{
\item{edge_list}{A dataframe containing a list of edges (see \code{\link{popsnetwork}}
for a description of possible formats).}
}
\value{
A list of ids of all source nodes in the network.
}
\description{
Get a list of source nodes from a network.
}
\details{
Extract a list of source nodes, i.e. nodes that do not have any
inputs, from a network.
}
\examples{
el <- data.frame(from=c("A", "B", "C"), to=c("C", "C", "D"))
sources(el)
}
